Fixes remove a file when run a program without output file

+++ gebrd/gebrd-job.c
@@ -1378,7 +1378,8 @@ static void job_assembly_cmdline(GebrdJob *job)
 		assemble_bc_cmd_line (expr_buf, ini, step);
 		prefix = g_strdup_printf("for (( counter=0; counter<%d; counter++ ))\ndo\n%s\n%s",
 					 counter, expr_buf->str, str_buf->str);
-		if(!gebr_geoxml_flow_io_get_output_append(job->flow) && !use_iter) {
+		if(!gebr_geoxml_flow_io_get_output_append(job->flow) && !use_iter &&
+		   strlen(gebr_geoxml_flow_io_get_output(job->flow)) > 0) {
 			remove = g_strdup_printf("\n\ttest $counter -eq 0 && rm -f %s\n", stdout_parsed);
 			g_string_prepend(job->parent.cmd_line, remove);
 			g_free(remove);
